requirements.txt
numpy
konlpy
python-telegram-bot~=11.1.0
